+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- Create project when unpacking ixp packages
+
 # v0.38.5
 - Resume timeouts on error